Move env schema into shared package
