+++ packages/worker/src/app/env.ts
@@ -1,5 +1,5 @@
 import { parseSafe } from 'remix/data-schema'
-import { EnvSchema, type AppEnv } from '../../../../types/env-schema.ts'
+import { EnvSchema, type AppEnv } from '@kody-internal/shared/env-schema.ts'
 
 export function getEnv(env: Env): AppEnv {
 	const result = parseSafe(EnvSchema, env)
